[MO][OPS-12140] Refactoring VerifyingYourBankAccountController
